pyproject.toml
[project]
name = "trapper-client"
version = "0.1.0"
description = "Python wrapper for Trapper API"
authors = [{ name = "ijfviana", email = "ijfviana@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12,<3.14"

dependencies = [
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "attrs>=25.3.0",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pyright>=1.1.407",
    "babel>=2.17.0",
    "sphinx>=8.2.3",
    "sphinx-autodoc-typehints>=3.5.2",
    "sphinx-rtd-theme>=3.0.2",
    "recommonmark>=0.7.1",
    "ruff>=0.14.3",
    "sphinx-autobuild>=2025.8.25",
    "sphinx-copybutton>=0.5.2",
]

[project.optional-dependencies]
typer = [
    "typer>=0.19.2",
    "typer-config>=1.4.2",
]

[project.scripts]
trapper-client = "trapper_client.ui.typer.main:app"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[tool.uv]
package = true

[tool.ruff]
# Set the maximum line length to 79.
line-length = 120

[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set.
extend-select = ["E501"]
